fix(cmake): support ARM64 kernel headers for musl builds
The musl build was failing on ARM64 because the kernel headers copy logic
only checked for x86_64-linux-gnu/asm but not aarch64-linux-gnu/asm.

Added architecture detection to find the correct asm headers path:
- aarch64/arm64 -> /usr/include/aarch64-linux-gnu/asm
- x86_64/amd64 -> /usr/include/x86_64-linux-gnu/asm
- i686/i386 -> /usr/include/i386-linux-gnu/asm
- Fallback: symlink asm -> asm-generic

+++ lib/platform/posix/terminal.c
@@ -603,7 +603,7 @@ terminal_capabilities_t apply_color_mode_override(terminal_capabilities_t caps) 
   if (opt_color_mode == COLOR_MODE_AUTO && platform_getenv("CLAUDECODE")) {
     log_debug("CLAUDECODE detected: forcing mono color mode");
     caps.color_level = TERM_COLOR_NONE;
-    caps.capabilities &= ~(TERM_CAP_COLOR_16 | TERM_CAP_COLOR_256 | TERM_CAP_COLOR_TRUE);
+    caps.capabilities &= ~(uint32_t)(TERM_CAP_COLOR_16 | TERM_CAP_COLOR_256 | TERM_CAP_COLOR_TRUE);
     caps.color_count = 0;
     return caps;
   }
